refactor: [WC-55] hook, page 일부 ts전환
