+++ src/pages/GuidePage/GuidePage.stories.tsx
@@ -1,0 +1,9 @@
+import { GuidePage } from '@/pages';
+
+export default {
+  title: 'Page/GuidePage',
+};
+
+export const Default = () => {
+  return <GuidePage />;
+};
