+++ src/pages/LoginPage/index.tsx
@@ -1,103 +1,91 @@
 import React from 'react';
 import Common from '@/styles';
 import styled from '@emotion/styled';
+import Swal from 'sweetalert2';
 import { useForm } from '@/hooks';
 import { useNavigate } from 'react-router-dom';
+import { useUser } from '@/hooks';
 import { Text, Button, Input, BackButton, Spinner } from '@/components';
 import {
   validateEmailEmpty,
   validateEmailForm,
-  validateNameEmpty,
-  validateNameLength,
   validatePasswordEmpty,
   validatePasswordLength,
-  validatePasswordConfirm,
 } from '@/validators';
-import { authApi } from '@/apis';
-import Swal from 'sweetalert2';
 
-const SignUpPage = () => {
+const LoginPage = () => {
+  const { login } = useUser();
   const navigate = useNavigate();
   const { isLoading, errors, handleChange, handleSubmit } = useForm({
     initialValues: {
       email: '',
-      userName: '',
       password: '',
-      passwordConfirm: '',
     },
-    onSubmit: async ({ email, userName, password }) => {
+    onSubmit: async ({ email, password }) => {
       try {
-        await authApi.signup({
-          name: userName,
-          email,
-          password,
-        });
+        await login(email, password);
 
         Swal.fire({
-          title: '🎉',
-          text: '환영합니다! 이제 로그인을 해주세요!',
+          title: '🥳',
+          text: '로그인 되었습니다!',
           confirmButtonColor: Common.colors.point,
         }).then(() => {
-          navigate('/login');
+          navigate('/', { replace: true });
         });
-      } catch (error) {
+      } catch (error: any) {
         Swal.fire({
           title: '🥲',
-          text: error,
+          text: error.message,
           confirmButtonColor: Common.colors.point,
         });
       }
     },
-    validate: ({ email, userName, password, passwordConfirm }) => {
-      const errors = {};
+    validate: ({ email, password }) => {
+      const errors: { [key: string]: string } = {};
+
       if (!validateEmailForm(email)) {
         errors.email = '올바른 이메일을 입력해주세요.';
       }
       if (!validateEmailEmpty(email)) errors.email = '이메일을 입력해주세요.';
-      if (!validateNameEmpty(userName))
-        errors.userName = '이름을 입력해주세요.';
-      if (!validateNameLength(userName))
-        errors.userName = '이름을 10글자 이내로 작성해주세요.';
       if (!validatePasswordLength(password)) {
         errors.password = '비밀번호를 8자 이상 작성해주세요.';
       }
       if (!validatePasswordEmpty(password)) {
         errors.password = '비밀번호를 입력해주세요.';
       }
-      if (!validatePasswordConfirm(password, passwordConfirm)) {
-        errors.passwordConfirm = '비밀번호가 일치하는지 확인해주세요.';
-      }
 
       return errors;
     },
   });
+
+  const handleClickSignUpButton = () => {
+    navigate('/signup');
+  };
 
   return (
     <Container>
       <StyledBackButton />
       <ContentContainer onSubmit={handleSubmit}>
         <Header size={Common.fontSize.large}>
-          회원가입하고 와플카드 만들러가요!
+          와플카드 대화에 참여해보시겠어요?
         </Header>
         <InputWrapper>
           <StyledText>이메일</StyledText>
           <Input name="email" type="email" onChange={handleChange} />
           <StyledText color="red">{errors.email}&nbsp;</StyledText>
-          <StyledText>이름(닉네임)</StyledText>
-          <Input name="userName" type="text" onChange={handleChange} />
-          <StyledText color="red">{errors.userName}&nbsp;</StyledText>
           <StyledText>비밀번호</StyledText>
           <Input name="password" type="password" onChange={handleChange} />
           <StyledText color="red">{errors.password}&nbsp;</StyledText>
-          <StyledText>비밀번호 확인</StyledText>
-          <Input
-            name="passwordConfirm"
-            type="password"
-            onChange={handleChange}
-          />
         </InputWrapper>
-        <StyledText color="red">{errors.passwordConfirm}&nbsp;</StyledText>
-        <StyledButton type="submit">가입하기</StyledButton>
+        <StyledButton type="submit">입장하기</StyledButton>
+        <StyledButton
+          type="button"
+          onClick={handleClickSignUpButton}
+          backgroundColor="white"
+          fontColor={Common.colors.button_font_dark}
+        >
+          가입하기
+        </StyledButton>
       </ContentContainer>
       <Spinner loading={isLoading} />
     </Container>
@@ -118,8 +106,11 @@ const Container = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  height: calc(100vh + 110px);
+  height: calc(100vh - 110px);
   padding: 100px 0;
+  @media ${Common.media.sm} {
+    height: calc(100vh - 60px);
+  }
 `;
 
 const ContentContainer = styled.form`
@@ -149,8 +140,11 @@ const Header = styled(Text)`
 
 const StyledButton = styled(Button)`
   width: 100%;
-  margin: 20px 0;
+  margin: 16px 0;
   font-weight: ${Common.fontWeight.bold};
+  &:nth-of-type(1) {
+    margin-top: 32px;
+  }
   @media ${Common.media.sm} {
     height: 40px;
     border-radius: 12px;
@@ -163,10 +157,6 @@ const StyledButton = styled(Button)`
 
 const StyledText = styled(Text)`
   margin: 16px 0;
-  @media ${Common.media.sm} {
-    font-size: ${Common.fontSize.small};
-    margin-bottom: 20px;
-  }
 `;
 
-export default SignUpPage;
+export default LoginPage;
